perf: split comment-section into isolated input components
Extracted CommentReplyInput and CommentComposeInput from
CommentSection so typing in inputs only re-renders the input
component — not the entire comment list.

Before: typing re-rendered all comment cards, drawer header,
bubble menu (CommentSection selected reply/comment from store)

After: only the active input component re-renders
