+++ package.json
@@ -2,7 +2,7 @@
   "name": "@fileverse-dev/ddoc",
   "private": false,
   "description": "DDoc",
-  "version": "3.0.96-zustand-2",
+  "version": "3.0.96-zustand-3",
   "main": "dist/index.es.js",
   "module": "dist/index.es.js",
   "exports": {
